fix(decode): return database status code and bad request for invalid payloads

+++ src/api/decode.h
@@ -2,5 +2,6 @@
 
 #include "uplink.h"
 #include <sqlite3.h>
+#include <stdint.h>
 
-int decode(sqlite3 *database, uplink_t *uplink);
+uint16_t decode(sqlite3 *database, uplink_t *uplink);
